intermediate check-in for dockerization of cores and server. Fighting my way through explicit and silent docker failures.
This commit deals with TC core dockerization. Some steps are working in some conditions, but end-to-end is not there yet.

Please note that the change to ehcache.xml either requires the cache folder to be /usr/local/transitclock/cache, or a soft link from your actual cache folder to that location.