Introduce the notion of a core cluster. The entire transit clock system has one or more such clusters, depending on the number of agencies that we are providing service for.
Each cluster has an instance that hosts java cores and an instance that
hosts postgres DBs. Each agency is running on a core, each core has a
java core and a DB. The tc-core-admin (hosted in a separate repo)
adminstrates a cluster, letting the admin add and remove cores and also
fielding status queries.

This admin tool has undegone local testing, but we need to create a
docker image and a gcloud deploy script for it and then test in the
actual glcoud context.